+++ LojasMoveis/LojasMoveis/Areas/Admin/Controllers/AdminPedidosController.cs
@@ -174,8 +174,11 @@ namespace LojasMoveis.Areas.Admin.Controllers
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var pedido = await _context.Pedidos.FindAsync(id);
-            _context.Pedidos.Remove(pedido);
-            await _context.SaveChangesAsync();
+            if (pedido != null)
+            {
+                _context.Pedidos.Remove(pedido);
+                await _context.SaveChangesAsync();
+            }
             return RedirectToAction(nameof(Index));
         }
